fix(server): move exceptions logging to separate filter
