+++ src/tests/Tochka.JsonRpc.Server.Tests/Filters/JsonRpcExceptionLoggingFilterTests.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using FluentAssertions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Abstractions;
@@ -13,68 +12,25 @@ using Moq;
 using NUnit.Framework;
 using Tochka.JsonRpc.Common.Features;
 using Tochka.JsonRpc.Common.Models.Request;
-using Tochka.JsonRpc.Common.Models.Response.Errors;
 using Tochka.JsonRpc.Server.Filters;
-using Tochka.JsonRpc.Server.Services;
 using Tochka.JsonRpc.Server.Settings;
 
 namespace Tochka.JsonRpc.Server.Tests.Filters;
 
 [TestFixture]
-internal class JsonRpcExceptionFilterTests
+public class JsonRpcExceptionLoggingFilterTests
 {
-    private Mock<IJsonRpcErrorFactory> errorFactoryMock;
     private JsonRpcServerOptions options;
-    private Mock<ILogger<JsonRpcExceptionFilter>> logMock;
-    private JsonRpcExceptionFilter exceptionFilter;
+    private Mock<ILogger<JsonRpcExceptionLoggingFilter>> logMock;
+    private JsonRpcExceptionLoggingFilter exceptionLoggingFilter;
 
     [SetUp]
     public void Setup()
     {
-        errorFactoryMock = new Mock<IJsonRpcErrorFactory>();
-        logMock = new Mock<ILogger<JsonRpcExceptionFilter>>();
         options = new JsonRpcServerOptions();
+        logMock = new Mock<ILogger<JsonRpcExceptionLoggingFilter>>();
 
-        exceptionFilter = new JsonRpcExceptionFilter(errorFactoryMock.Object, Options.Create(options), logMock.Object);
-    }
-
-    [Test]
-    public void OnException_NotJsonRpc_DoNothing()
-    {
-        var exception = new ArgumentException();
-        var httpContext = new DefaultHttpContext();
-        var actionContext = new ActionContext(httpContext, new RouteData(), new ActionDescriptor(), new ModelStateDictionary());
-        var context = new ExceptionContext(actionContext, new List<IFilterMetadata>())
-        {
-            Exception = exception
-        };
-
-        exceptionFilter.OnException(context);
-
-        context.Result.Should().BeNull();
-    }
-
-    [Test]
-    public void OnException_JsonRpcCall_SetErrorResult()
-    {
-        var exception = new ArgumentException();
-        var httpContext = new DefaultHttpContext();
-        httpContext.Features.Set<IJsonRpcFeature>(new JsonRpcFeature { Call = Mock.Of<ICall>() });
-        var actionContext = new ActionContext(httpContext, new RouteData(), new ActionDescriptor(), new ModelStateDictionary());
-        var context = new ExceptionContext(actionContext, new List<IFilterMetadata>())
-        {
-            Exception = exception
-        };
-        var error = Mock.Of<IError>();
-        errorFactoryMock.Setup(f => f.Exception(exception))
-            .Returns(error)
-            .Verifiable();
-
-        exceptionFilter.OnException(context);
-
-        var expected = new ObjectResult(error);
-        context.Result.Should().BeEquivalentTo(expected);
-        errorFactoryMock.Verify();
+        exceptionLoggingFilter = new JsonRpcExceptionLoggingFilter(Options.Create(options), logMock.Object);
     }
 
     [Test]
@@ -92,7 +48,7 @@ internal class JsonRpcExceptionFilterTests
         logMock.Setup(l => l.Log(LogLevel.Error, It.IsAny<EventId>(), It.IsAny<It.IsAnyType>(), exception, It.IsAny<Func<It.IsAnyType, Exception?, string>>()))
             .Verifiable();
 
-        exceptionFilter.OnException(context);
+        exceptionLoggingFilter.OnException(context);
 
         logMock.Verify();
     }
@@ -110,7 +66,7 @@ internal class JsonRpcExceptionFilterTests
         };
         options.LogExceptions = false;
 
-        exceptionFilter.OnException(context);
+        exceptionLoggingFilter.OnException(context);
 
         logMock.VerifyNoOtherCalls();
     }
